Refactor Notebook and Store classes to streamline notebook creation and retrieval logic. Introduce create_notebook_node functions for improved handling of different file types. Update tests to reflect changes in notebook initialization and language detection.
