+++ src/nbstore/parsers/markdown.py
@@ -2,11 +2,10 @@ from __future__ import annotations
 
 import re
 from dataclasses import dataclass
-from functools import cache
 from typing import TYPE_CHECKING, ClassVar
 
 if TYPE_CHECKING:
-    from collections.abc import Iterator
+    from collections.abc import Iterable, Iterator
     from typing import Self
 
 
@@ -301,8 +300,7 @@ def iter_elements(
             yield from iter_elements(text, elem[0], elem[1], classes[1:], url)
 
 
-@cache
-def get_language(text: str) -> str | None:
+def get_language(elems: Iterable[str | Element]) -> str | None:
     """Get the language of the first code block in the text.
 
     If there is no code block for a Jupyter notebook, return None.
@@ -317,11 +315,11 @@ def get_language(text: str) -> str | None:
     languages = {}
     identifiers = []
 
-    for elem in iter_elements(text):
+    for elem in elems:
         if isinstance(elem, CodeBlock) and elem.identifier and elem.classes:
             language = elem.classes[0].removeprefix(".")
             languages[elem.identifier] = language
-        elif isinstance(elem, Image) and elem.identifier and elem.url in (".md", ""):
+        elif isinstance(elem, Image) and elem.identifier and elem.url in (".md", "."):
             if elem.identifier in languages:
                 return languages[elem.identifier]
             identifiers.append(elem.identifier)
